requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
ldap3==2.5.1
pyjwt==2.3.0